requirements.txt
pycryptodomex>=3.19.0
mmh3>=4.0.0
